+++ spec/station12/movies_controller_spec.rb
